split tests + add protocols to prot.conf
