+++ xmipptomo/protocols.conf
@@ -13,11 +13,14 @@ Tomography = [
 		{"tag": "protocol_group", "text": "Extract", "openItem": "False", "children": []}
     ]},
     {"tag": "section", "text": "Preprocessing", "children": []},
-    {"tag": "section", "text": "Subtomogram averaging", "children": []},
+    {"tag": "section", "text": "Subtomogram averaging", "children": [
+        {"tag": "protocol", "value": "XmippProtCLTomo",   "text": "default"}
+        ]},
     {"tag": "section", "text": "Postprocessing", "children": [
         {"tag": "protocol", "value": "XmippProtApplyTransformSubtomo", "text": "default"},
         {"tag": "protocol", "value": "XmippProtUndoAlignSubtomo", "text": "default"},
         {"tag": "protocol", "value": "XmippProtSubtomoMapBack", "text": "default"},
+        {"tag": "protocol", "value": "XmippProtMonoTomo", "text": "default"},
         {"tag": "protocol_group", "text": "Denoise", "openItem": "False", "children": []},
 	    {"tag": "protocol_group", "text": "Segmentation", "openItem": "False", "children": []}
     ]}]
